added services and controlleres for user,admin,trainer,subject,topic and pdf

+++ edunaova_spring/demo/src/main/java/com/example/demo/repo/AdminRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entities.Admin;
 
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 //jpa repo replaces dao layer
 }
